simpler maven configuring script
