+++ buildSrc/build.gradle.kts
@@ -11,8 +11,6 @@ repositories {
 }
 
 dependencies {
-    api(gradleApi())
-    api(gradleKotlinDsl())
     implementation(kotlin("stdlib-jdk8"))
 }
